requirements.txt
baku-config>=1.0.0
pyzmq>=25.0.0

